Reduce codeclimate complexity
