+++ app/src/main/java/com/github/sdp/mediato/data/Utils.java
@@ -1,6 +1,15 @@
 package com.github.sdp.mediato.data;
 
+import com.github.sdp.mediato.model.Location;
+import com.github.sdp.mediato.model.User;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.ValueEventListener;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
 
 /**
  * This is a Utils class for the database classes
@@ -33,5 +42,37 @@ class Utils {
      */
     static DatabaseReference getCollectionReference(String username, String collectionName) {
         return CollectionsDatabase.database.getReference().child(USERS_PATH + username + USER_COLLECTIONS_PATH + collectionName);
+    }
+
+    /**
+     * Helper method for UserDatabase that returns the nearby users and completes the future
+     *
+     * @param future to track the state
+     * @param location center location of the current user
+     * @param username username of the current user
+     * @return a list of strings with the users' usernames
+     */
+    static List<String> findNearbyUsers(CompletableFuture<List<String>> future, Location location, String username) {
+        List<String> nearbyUsers = new ArrayList<>();
+        UserDatabase.database.getReference(USERS_PATH).addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                dataSnapshot.getChildren().forEach(
+                        userSnapshot -> {
+                            User user = userSnapshot.getValue(User.class);
+                            if (!userSnapshot.getKey().equals(username) && user.getLocation().isValid() && user.getLocation().isInRadius(location)) {
+                                nearbyUsers.add(userSnapshot.getKey());
+                            }
+                        }
+                );
+                future.complete(nearbyUsers);
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+                future.completeExceptionally(databaseError.toException());
+            }
+        });
+        return nearbyUsers;
     }
 }
